Moved HMI RPCs.
