+++ src/components/application_manager/rpc_plugins/vehicle_info_plugin/include/vehicle_info_plugin/commands/hmi/on_vi_vehicle_data_notification.h
@@ -30,36 +30,40 @@
  * POSSIBILITY OF SUCH DAMAGE.
  */
 
-#ifndef SRC_COMPONENTS_APPLICATION_MANAGER_RPC_PLUGINS_SDL_RPC_PLUGIN_INCLUDE_SDL_RPC_PLUGIN_COMMANDS_HMI_VI_READ_DID_REQUEST_H_
-#define SRC_COMPONENTS_APPLICATION_MANAGER_RPC_PLUGINS_SDL_RPC_PLUGIN_INCLUDE_SDL_RPC_PLUGIN_COMMANDS_HMI_VI_READ_DID_REQUEST_H_
+#ifndef SRC_COMPONENTS_APPLICATION_MANAGER_RPC_PLUGINS_SDL_RPC_PLUGIN_INCLUDE_SDL_RPC_PLUGIN_COMMANDS_HMI_ON_VI_VEHICLE_DATA_NOTIFICATION_H_
+#define SRC_COMPONENTS_APPLICATION_MANAGER_RPC_PLUGINS_SDL_RPC_PLUGIN_INCLUDE_SDL_RPC_PLUGIN_COMMANDS_HMI_ON_VI_VEHICLE_DATA_NOTIFICATION_H_
 
-#include "application_manager/commands/request_to_hmi.h"
+#include "application_manager/commands/notification_from_hmi.h"
+#include "application_manager/application_manager.h"
 
-namespace sdl_rpc_plugin {
+namespace vehicle_info_plugin {
 namespace app_mngr = application_manager;
 
 namespace commands {
 
 /**
- * @brief VIReadDIDRequest command class
+ * @brief OnVIVehicleDataNotification command class
+ * Sent by HMI for the periodic and non periodic vehicle data
  **/
-class VIReadDIDRequest : public app_mngr::commands::RequestToHMI {
+class OnVIVehicleDataNotification
+    : public app_mngr::commands::NotificationFromHMI {
  public:
   /**
-   * @brief VIReadDIDRequest class constructor
+   * @brief OnVIVehicleDataNotification class constructor
    *
    * @param message Incoming SmartObject message
    **/
-  VIReadDIDRequest(const app_mngr::commands::MessageSharedPtr& message,
-                   app_mngr::ApplicationManager& application_manager,
-                   app_mngr::rpc_service::RPCService& rpc_service,
-                   app_mngr::HMICapabilities& hmi_capabilities,
-                   policy::PolicyHandlerInterface& policy_handle);
+  OnVIVehicleDataNotification(
+      const app_mngr::commands::MessageSharedPtr& message,
+      app_mngr::ApplicationManager& application_manager,
+      app_mngr::rpc_service::RPCService& rpc_service,
+      app_mngr::HMICapabilities& hmi_capabilities,
+      policy::PolicyHandlerInterface& policy_handle);
 
   /**
-   * @brief VIReadDIDRequest class destructor
+   * @brief OnVIVehicleDataNotification class destructor
    **/
-  virtual ~VIReadDIDRequest();
+  virtual ~OnVIVehicleDataNotification();
 
   /**
    * @brief Execute command
@@ -67,11 +71,11 @@ class VIReadDIDRequest : public app_mngr::commands::RequestToHMI {
   virtual void Run();
 
  private:
-  DISALLOW_COPY_AND_ASSIGN(VIReadDIDRequest);
+  DISALLOW_COPY_AND_ASSIGN(OnVIVehicleDataNotification);
 };
 
 }  // namespace commands
 
 }  // namespace application_manager
 
-#endif  // SRC_COMPONENTS_APPLICATION_MANAGER_RPC_PLUGINS_SDL_RPC_PLUGIN_INCLUDE_SDL_RPC_PLUGIN_COMMANDS_HMI_VI_READ_DID_REQUEST_H_
+#endif  // SRC_COMPONENTS_APPLICATION_MANAGER_RPC_PLUGINS_SDL_RPC_PLUGIN_INCLUDE_SDL_RPC_PLUGIN_COMMANDS_HMI_ON_VI_VEHICLE_DATA_NOTIFICATION_H_
