+++ src/components/application_manager/rpc_plugins/vehicle_info_plugin/include/vehicle_info_plugin/commands/hmi/vi_diagnostic_message_request.h
@@ -30,36 +30,37 @@
  * POSSIBILITY OF SUCH DAMAGE.
  */
 
-#ifndef SRC_COMPONENTS_APPLICATION_MANAGER_RPC_PLUGINS_SDL_RPC_PLUGIN_INCLUDE_SDL_RPC_PLUGIN_COMMANDS_HMI_VI_GET_DTCS_REQUEST_H_
-#define SRC_COMPONENTS_APPLICATION_MANAGER_RPC_PLUGINS_SDL_RPC_PLUGIN_INCLUDE_SDL_RPC_PLUGIN_COMMANDS_HMI_VI_GET_DTCS_REQUEST_H_
+#ifndef SRC_COMPONENTS_APPLICATION_MANAGER_RPC_PLUGINS_SDL_RPC_PLUGIN_INCLUDE_SDL_RPC_PLUGIN_COMMANDS_HMI_VI_DIAGNOSTIC_MESSAGE_REQUEST_H_
+#define SRC_COMPONENTS_APPLICATION_MANAGER_RPC_PLUGINS_SDL_RPC_PLUGIN_INCLUDE_SDL_RPC_PLUGIN_COMMANDS_HMI_VI_DIAGNOSTIC_MESSAGE_REQUEST_H_
 
 #include "application_manager/commands/request_to_hmi.h"
 
-namespace sdl_rpc_plugin {
+namespace vehicle_info_plugin {
 namespace app_mngr = application_manager;
 
 namespace commands {
 
 /**
- * @brief VIGetDTCsRequest command class
+ * @brief VIDiagnosticMessageRequest command class
  **/
-class VIGetDTCsRequest : public app_mngr::commands::RequestToHMI {
+class VIDiagnosticMessageRequest : public app_mngr::commands::RequestToHMI {
  public:
   /**
-   * @brief VIGetDTCsRequest class constructor
+   * @brief VIDiagnosticMessageRequest class constructor
    *
    * @param message Incoming SmartObject message
    **/
-  VIGetDTCsRequest(const app_mngr::commands::MessageSharedPtr& message,
-                   app_mngr::ApplicationManager& application_manager,
-                   app_mngr::rpc_service::RPCService& rpc_service,
-                   app_mngr::HMICapabilities& hmi_capabilities,
-                   policy::PolicyHandlerInterface& policy_handle);
+  VIDiagnosticMessageRequest(
+      const app_mngr::commands::MessageSharedPtr& message,
+      app_mngr::ApplicationManager& application_manager,
+      app_mngr::rpc_service::RPCService& rpc_service,
+      app_mngr::HMICapabilities& hmi_capabilities,
+      policy::PolicyHandlerInterface& policy_handle);
 
   /**
-   * @brief VIGetDTCsRequest class destructor
+   * @brief VIDiagnosticMessageRequest class destructor
    **/
-  virtual ~VIGetDTCsRequest();
+  virtual ~VIDiagnosticMessageRequest();
 
   /**
    * @brief Execute command
@@ -67,11 +68,11 @@ class VIGetDTCsRequest : public app_mngr::commands::RequestToHMI {
   virtual void Run();
 
  private:
-  DISALLOW_COPY_AND_ASSIGN(VIGetDTCsRequest);
+  DISALLOW_COPY_AND_ASSIGN(VIDiagnosticMessageRequest);
 };
 
 }  // namespace commands
 
 }  // namespace application_manager
 
-#endif  // SRC_COMPONENTS_APPLICATION_MANAGER_RPC_PLUGINS_SDL_RPC_PLUGIN_INCLUDE_SDL_RPC_PLUGIN_COMMANDS_HMI_VI_GET_DTCS_REQUEST_H_
+#endif  // SRC_COMPONENTS_APPLICATION_MANAGER_RPC_PLUGINS_SDL_RPC_PLUGIN_INCLUDE_SDL_RPC_PLUGIN_COMMANDS_HMI_VI_DIAGNOSTIC_MESSAGE_REQUEST_H_
