+++ src/components/application_manager/rpc_plugins/vehicle_info_plugin/include/vehicle_info_plugin/commands/hmi/vi_get_dtcs_response.h
@@ -30,38 +30,36 @@
  * POSSIBILITY OF SUCH DAMAGE.
  */
 
-#ifndef SRC_COMPONENTS_APPLICATION_MANAGER_RPC_PLUGINS_SDL_RPC_PLUGIN_INCLUDE_SDL_RPC_PLUGIN_COMMANDS_HMI_VI_SUBSCRIBE_VEHICLE_DATA_RESPONSE_H_
-#define SRC_COMPONENTS_APPLICATION_MANAGER_RPC_PLUGINS_SDL_RPC_PLUGIN_INCLUDE_SDL_RPC_PLUGIN_COMMANDS_HMI_VI_SUBSCRIBE_VEHICLE_DATA_RESPONSE_H_
+#ifndef SRC_COMPONENTS_APPLICATION_MANAGER_RPC_PLUGINS_SDL_RPC_PLUGIN_INCLUDE_SDL_RPC_PLUGIN_COMMANDS_HMI_VI_GET_DTCS_RESPONSE_H_
+#define SRC_COMPONENTS_APPLICATION_MANAGER_RPC_PLUGINS_SDL_RPC_PLUGIN_INCLUDE_SDL_RPC_PLUGIN_COMMANDS_HMI_VI_GET_DTCS_RESPONSE_H_
 
 #include "application_manager/commands/response_from_hmi.h"
 
-namespace sdl_rpc_plugin {
+namespace vehicle_info_plugin {
 namespace app_mngr = application_manager;
 
 namespace commands {
 
 /**
- * @brief VISubscribeVehicleDataResponse command class
+ * @brief VIGetDTCsResponse command class
  **/
-class VISubscribeVehicleDataResponse
-    : public app_mngr::commands::ResponseFromHMI {
+class VIGetDTCsResponse : public app_mngr::commands::ResponseFromHMI {
  public:
   /**
-   * @brief VISubscribeVehicleDataResponse class constructor
+   * @brief VIGetDTCsResponse class constructor
    *
    * @param message Incoming SmartObject message
    **/
-  VISubscribeVehicleDataResponse(
-      const app_mngr::commands::MessageSharedPtr& message,
-      app_mngr::ApplicationManager& application_manager,
-      app_mngr::rpc_service::RPCService& rpc_service,
-      app_mngr::HMICapabilities& hmi_capabilities,
-      policy::PolicyHandlerInterface& policy_handle);
+  VIGetDTCsResponse(const app_mngr::commands::MessageSharedPtr& message,
+                    app_mngr::ApplicationManager& application_manager,
+                    app_mngr::rpc_service::RPCService& rpc_service,
+                    app_mngr::HMICapabilities& hmi_capabilities,
+                    policy::PolicyHandlerInterface& policy_handle);
 
   /**
-   * @brief VISubscribeVehicleDataResponse class destructor
+   * @brief VIGetDTCsResponse class destructor
    **/
-  virtual ~VISubscribeVehicleDataResponse();
+  virtual ~VIGetDTCsResponse();
 
   /**
    * @brief Execute command
@@ -69,11 +67,11 @@ class VISubscribeVehicleDataResponse
   virtual void Run();
 
  private:
-  DISALLOW_COPY_AND_ASSIGN(VISubscribeVehicleDataResponse);
+  DISALLOW_COPY_AND_ASSIGN(VIGetDTCsResponse);
 };
 
 }  // namespace commands
 
 }  // namespace application_manager
 
-#endif  // SRC_COMPONENTS_APPLICATION_MANAGER_RPC_PLUGINS_SDL_RPC_PLUGIN_INCLUDE_SDL_RPC_PLUGIN_COMMANDS_HMI_VI_SUBSCRIBE_VEHICLE_DATA_RESPONSE_H_
+#endif  // SRC_COMPONENTS_APPLICATION_MANAGER_RPC_PLUGINS_SDL_RPC_PLUGIN_INCLUDE_SDL_RPC_PLUGIN_COMMANDS_HMI_VI_GET_DTCS_RESPONSE_H_
