+++ src/components/application_manager/rpc_plugins/vehicle_info_plugin/include/vehicle_info_plugin/commands/hmi/vi_read_did_request.h
@@ -30,38 +30,36 @@
  * POSSIBILITY OF SUCH DAMAGE.
  */
 
-#ifndef SRC_COMPONENTS_APPLICATION_MANAGER_RPC_PLUGINS_SDL_RPC_PLUGIN_INCLUDE_SDL_RPC_PLUGIN_COMMANDS_HMI_VI_UNSUBSCRIBE_VEHICLE_DATA_RESPONSE_H_
-#define SRC_COMPONENTS_APPLICATION_MANAGER_RPC_PLUGINS_SDL_RPC_PLUGIN_INCLUDE_SDL_RPC_PLUGIN_COMMANDS_HMI_VI_UNSUBSCRIBE_VEHICLE_DATA_RESPONSE_H_
+#ifndef SRC_COMPONENTS_APPLICATION_MANAGER_RPC_PLUGINS_SDL_RPC_PLUGIN_INCLUDE_SDL_RPC_PLUGIN_COMMANDS_HMI_VI_READ_DID_REQUEST_H_
+#define SRC_COMPONENTS_APPLICATION_MANAGER_RPC_PLUGINS_SDL_RPC_PLUGIN_INCLUDE_SDL_RPC_PLUGIN_COMMANDS_HMI_VI_READ_DID_REQUEST_H_
 
-#include "application_manager/commands/response_from_hmi.h"
+#include "application_manager/commands/request_to_hmi.h"
 
-namespace sdl_rpc_plugin {
+namespace vehicle_info_plugin {
 namespace app_mngr = application_manager;
 
 namespace commands {
 
 /**
- * @brief VIUnsubscribeVehicleDataResponse command class
+ * @brief VIReadDIDRequest command class
  **/
-class VIUnsubscribeVehicleDataResponse
-    : public app_mngr::commands::ResponseFromHMI {
+class VIReadDIDRequest : public app_mngr::commands::RequestToHMI {
  public:
   /**
-   * @brief VIUnsubscribeVehicleDataResponse class constructor
+   * @brief VIReadDIDRequest class constructor
    *
    * @param message Incoming SmartObject message
    **/
-  VIUnsubscribeVehicleDataResponse(
-      const app_mngr::commands::MessageSharedPtr& message,
-      app_mngr::ApplicationManager& application_manager,
-      app_mngr::rpc_service::RPCService& rpc_service,
-      app_mngr::HMICapabilities& hmi_capabilities,
-      policy::PolicyHandlerInterface& policy_handle);
+  VIReadDIDRequest(const app_mngr::commands::MessageSharedPtr& message,
+                   app_mngr::ApplicationManager& application_manager,
+                   app_mngr::rpc_service::RPCService& rpc_service,
+                   app_mngr::HMICapabilities& hmi_capabilities,
+                   policy::PolicyHandlerInterface& policy_handle);
 
   /**
-   * @brief VIUnsubscribeVehicleDataResponse class destructor
+   * @brief VIReadDIDRequest class destructor
    **/
-  virtual ~VIUnsubscribeVehicleDataResponse();
+  virtual ~VIReadDIDRequest();
 
   /**
    * @brief Execute command
@@ -69,11 +67,11 @@ class VIUnsubscribeVehicleDataResponse
   virtual void Run();
 
  private:
-  DISALLOW_COPY_AND_ASSIGN(VIUnsubscribeVehicleDataResponse);
+  DISALLOW_COPY_AND_ASSIGN(VIReadDIDRequest);
 };
 
 }  // namespace commands
 
 }  // namespace application_manager
 
-#endif  // SRC_COMPONENTS_APPLICATION_MANAGER_RPC_PLUGINS_SDL_RPC_PLUGIN_INCLUDE_SDL_RPC_PLUGIN_COMMANDS_HMI_VI_UNSUBSCRIBE_VEHICLE_DATA_RESPONSE_H_
+#endif  // SRC_COMPONENTS_APPLICATION_MANAGER_RPC_PLUGINS_SDL_RPC_PLUGIN_INCLUDE_SDL_RPC_PLUGIN_COMMANDS_HMI_VI_READ_DID_REQUEST_H_
