+++ src/components/application_manager/rpc_plugins/vehicle_info_plugin/include/vehicle_info_plugin/commands/hmi/vi_read_did_response.h
@@ -29,35 +29,49 @@
  * ARISING IN ANY WAY OUT OF THE USE OF THIS SOFTWARE, EVEN IF ADVISED OF THE
  * POSSIBILITY OF SUCH DAMAGE.
  */
-#include "sdl_rpc_plugin/commands/hmi/vi_unsubscribe_vehicle_data_response.h"
-#include "application_manager/event_engine/event.h"
-#include "interfaces/MOBILE_API.h"
 
-namespace sdl_rpc_plugin {
-using namespace application_manager;
+#ifndef SRC_COMPONENTS_APPLICATION_MANAGER_RPC_PLUGINS_SDL_RPC_PLUGIN_INCLUDE_SDL_RPC_PLUGIN_COMMANDS_HMI_VI_READ_DID_RESPONSE_H_
+#define SRC_COMPONENTS_APPLICATION_MANAGER_RPC_PLUGINS_SDL_RPC_PLUGIN_INCLUDE_SDL_RPC_PLUGIN_COMMANDS_HMI_VI_READ_DID_RESPONSE_H_
+
+#include "application_manager/commands/response_from_hmi.h"
+
+namespace vehicle_info_plugin {
+namespace app_mngr = application_manager;
+
 namespace commands {
 
-VIUnsubscribeVehicleDataResponse::VIUnsubscribeVehicleDataResponse(
-    const application_manager::commands::MessageSharedPtr& message,
-    ApplicationManager& application_manager,
-    rpc_service::RPCService& rpc_service,
-    HMICapabilities& hmi_capabilities,
-    policy::PolicyHandlerInterface& policy_handle)
-    : ResponseFromHMI(message,
-                      application_manager,
-                      rpc_service,
-                      hmi_capabilities,
-                      policy_handle) {}
+/**
+ * @brief VIReadDIDResponse command class
+ **/
+class VIReadDIDResponse : public app_mngr::commands::ResponseFromHMI {
+ public:
+  /**
+   * @brief VIReadDIDResponse class constructor
+   *
+   * @param message Incoming SmartObject message
+   **/
+  VIReadDIDResponse(const app_mngr::commands::MessageSharedPtr& message,
+                    app_mngr::ApplicationManager& application_manager,
+                    app_mngr::rpc_service::RPCService& rpc_service,
+                    app_mngr::HMICapabilities& hmi_capabilities,
+                    policy::PolicyHandlerInterface& policy_handle);
 
-VIUnsubscribeVehicleDataResponse::~VIUnsubscribeVehicleDataResponse() {}
+  /**
+   * @brief VIReadDIDResponse class destructor
+   **/
+  virtual ~VIReadDIDResponse();
 
-void VIUnsubscribeVehicleDataResponse::Run() {
-  LOG4CXX_AUTO_TRACE(logger_);
-  event_engine::Event event(
-      hmi_apis::FunctionID::VehicleInfo_UnsubscribeVehicleData);
-  event.set_smart_object(*message_);
-  event.raise(application_manager_.event_dispatcher());
-}
+  /**
+   * @brief Execute command
+   **/
+  virtual void Run();
+
+ private:
+  DISALLOW_COPY_AND_ASSIGN(VIReadDIDResponse);
+};
 
 }  // namespace commands
+
 }  // namespace application_manager
+
+#endif  // SRC_COMPONENTS_APPLICATION_MANAGER_RPC_PLUGINS_SDL_RPC_PLUGIN_INCLUDE_SDL_RPC_PLUGIN_COMMANDS_HMI_VI_READ_DID_RESPONSE_H_
