+++ src/components/application_manager/rpc_plugins/vehicle_info_plugin/include/vehicle_info_plugin/commands/hmi/vi_subscribe_vehicle_data_request.h
@@ -30,46 +30,49 @@
  * POSSIBILITY OF SUCH DAMAGE.
  */
 
-#include "sdl_rpc_plugin/commands/hmi/on_vi_vehicle_data_notification.h"
-#include "interfaces/MOBILE_API.h"
+#ifndef SRC_COMPONENTS_APPLICATION_MANAGER_RPC_PLUGINS_SDL_RPC_PLUGIN_INCLUDE_SDL_RPC_PLUGIN_COMMANDS_HMI_VI_SUBSCRIBE_VEHICLE_DATA_REQUEST_H_
+#define SRC_COMPONENTS_APPLICATION_MANAGER_RPC_PLUGINS_SDL_RPC_PLUGIN_INCLUDE_SDL_RPC_PLUGIN_COMMANDS_HMI_VI_SUBSCRIBE_VEHICLE_DATA_REQUEST_H_
 
-namespace sdl_rpc_plugin {
-using namespace application_manager;
+#include "application_manager/commands/request_to_hmi.h"
+
+namespace vehicle_info_plugin {
+namespace app_mngr = application_manager;
 
 namespace commands {
 
-OnVIVehicleDataNotification::OnVIVehicleDataNotification(
-    const application_manager::commands::MessageSharedPtr& message,
-    ApplicationManager& application_manager,
-    rpc_service::RPCService& rpc_service,
-    HMICapabilities& hmi_capabilities,
-    policy::PolicyHandlerInterface& policy_handle)
-    : NotificationFromHMI(message,
-                          application_manager,
-                          rpc_service,
-                          hmi_capabilities,
-                          policy_handle) {}
+/**
+ * @brief VISubscribeVehicleDataRequest command class
+ **/
+class VISubscribeVehicleDataRequest : public app_mngr::commands::RequestToHMI {
+ public:
+  /**
+   * @brief VISubscribeVehicleDataRequest class constructor
+   *
+   * @param message Incoming SmartObject message
+   **/
+  VISubscribeVehicleDataRequest(
+      const app_mngr::commands::MessageSharedPtr& message,
+      app_mngr::ApplicationManager& application_manager,
+      app_mngr::rpc_service::RPCService& rpc_service,
+      app_mngr::HMICapabilities& hmi_capabilities,
+      policy::PolicyHandlerInterface& policy_handle);
 
-OnVIVehicleDataNotification::~OnVIVehicleDataNotification() {}
+  /**
+   * @brief VISubscribeVehicleDataRequest class destructor
+   **/
+  virtual ~VISubscribeVehicleDataRequest();
 
-void OnVIVehicleDataNotification::Run() {
-  LOG4CXX_AUTO_TRACE(logger_);
+  /**
+   * @brief Execute command
+   **/
+  virtual void Run();
 
-  // prepare SmartObject for mobile factory
-  (*message_)[strings::params][strings::function_id] =
-      static_cast<int32_t>(mobile_apis::FunctionID::eType::OnVehicleDataID);
-
-  const smart_objects::SmartObject& msg_params =
-      (*message_)[strings::msg_params];
-  if (msg_params.keyExists(strings::odometer)) {
-    application_manager_.IviInfoUpdated(
-        mobile_apis::VehicleDataType::VEHICLEDATA_ODOMETER,
-        msg_params[strings::odometer].asInt());
-  }
-
-  SendNotificationToMobile(message_);
-}
+ private:
+  DISALLOW_COPY_AND_ASSIGN(VISubscribeVehicleDataRequest);
+};
 
 }  // namespace commands
 
 }  // namespace application_manager
+
+#endif  // SRC_COMPONENTS_APPLICATION_MANAGER_RPC_PLUGINS_SDL_RPC_PLUGIN_INCLUDE_SDL_RPC_PLUGIN_COMMANDS_HMI_VI_SUBSCRIBE_VEHICLE_DATA_REQUEST_H_
