+++ src/components/application_manager/rpc_plugins/vehicle_info_plugin/include/vehicle_info_plugin/commands/hmi/vi_subscribe_vehicle_data_response.h
@@ -30,37 +30,38 @@
  * POSSIBILITY OF SUCH DAMAGE.
  */
 
-#ifndef SRC_COMPONENTS_APPLICATION_MANAGER_RPC_PLUGINS_SDL_RPC_PLUGIN_INCLUDE_SDL_RPC_PLUGIN_COMMANDS_HMI_VI_GET_VEHICLE_DATA_RESPONSE_H_
-#define SRC_COMPONENTS_APPLICATION_MANAGER_RPC_PLUGINS_SDL_RPC_PLUGIN_INCLUDE_SDL_RPC_PLUGIN_COMMANDS_HMI_VI_GET_VEHICLE_DATA_RESPONSE_H_
+#ifndef SRC_COMPONENTS_APPLICATION_MANAGER_RPC_PLUGINS_SDL_RPC_PLUGIN_INCLUDE_SDL_RPC_PLUGIN_COMMANDS_HMI_VI_SUBSCRIBE_VEHICLE_DATA_RESPONSE_H_
+#define SRC_COMPONENTS_APPLICATION_MANAGER_RPC_PLUGINS_SDL_RPC_PLUGIN_INCLUDE_SDL_RPC_PLUGIN_COMMANDS_HMI_VI_SUBSCRIBE_VEHICLE_DATA_RESPONSE_H_
 
 #include "application_manager/commands/response_from_hmi.h"
-#include "application_manager/application_manager.h"
 
-namespace sdl_rpc_plugin {
+namespace vehicle_info_plugin {
 namespace app_mngr = application_manager;
 
 namespace commands {
 
 /**
- * @brief VIGetVehicleDataResponse command class
+ * @brief VISubscribeVehicleDataResponse command class
  **/
-class VIGetVehicleDataResponse : public app_mngr::commands::ResponseFromHMI {
+class VISubscribeVehicleDataResponse
+    : public app_mngr::commands::ResponseFromHMI {
  public:
   /**
-   * @brief VIGetVehicleDataResponse class constructor
+   * @brief VISubscribeVehicleDataResponse class constructor
    *
    * @param message Incoming SmartObject message
    **/
-  VIGetVehicleDataResponse(const app_mngr::commands::MessageSharedPtr& message,
-                           app_mngr::ApplicationManager& application_manager,
-                           app_mngr::rpc_service::RPCService& rpc_service,
-                           app_mngr::HMICapabilities& hmi_capabilities,
-                           policy::PolicyHandlerInterface& policy_handle);
+  VISubscribeVehicleDataResponse(
+      const app_mngr::commands::MessageSharedPtr& message,
+      app_mngr::ApplicationManager& application_manager,
+      app_mngr::rpc_service::RPCService& rpc_service,
+      app_mngr::HMICapabilities& hmi_capabilities,
+      policy::PolicyHandlerInterface& policy_handle);
 
   /**
-   * @brief VIGetVehicleDataResponse class destructor
+   * @brief VISubscribeVehicleDataResponse class destructor
    **/
-  virtual ~VIGetVehicleDataResponse();
+  virtual ~VISubscribeVehicleDataResponse();
 
   /**
    * @brief Execute command
@@ -68,11 +69,11 @@ class VIGetVehicleDataResponse : public app_mngr::commands::ResponseFromHMI {
   virtual void Run();
 
  private:
-  DISALLOW_COPY_AND_ASSIGN(VIGetVehicleDataResponse);
+  DISALLOW_COPY_AND_ASSIGN(VISubscribeVehicleDataResponse);
 };
 
 }  // namespace commands
 
 }  // namespace application_manager
 
-#endif  // SRC_COMPONENTS_APPLICATION_MANAGER_RPC_PLUGINS_SDL_RPC_PLUGIN_INCLUDE_SDL_RPC_PLUGIN_COMMANDS_HMI_VI_GET_VEHICLE_DATA_RESPONSE_H_
+#endif  // SRC_COMPONENTS_APPLICATION_MANAGER_RPC_PLUGINS_SDL_RPC_PLUGIN_INCLUDE_SDL_RPC_PLUGIN_COMMANDS_HMI_VI_SUBSCRIBE_VEHICLE_DATA_RESPONSE_H_
