+++ src/components/application_manager/rpc_plugins/vehicle_info_plugin/src/commands/hmi/vi_get_dtcs_request.cc
@@ -30,14 +30,14 @@
  * POSSIBILITY OF SUCH DAMAGE.
  */
 
-#include "sdl_rpc_plugin/commands/hmi/vi_unsubscribe_vehicle_data_request.h"
+#include "vehicle_info_plugin/commands/hmi/vi_get_dtcs_request.h"
 
-namespace sdl_rpc_plugin {
+namespace vehicle_info_plugin {
 using namespace application_manager;
 
 namespace commands {
 
-VIUnsubscribeVehicleDataRequest::VIUnsubscribeVehicleDataRequest(
+VIGetDTCsRequest::VIGetDTCsRequest(
     const application_manager::commands::MessageSharedPtr& message,
     ApplicationManager& application_manager,
     rpc_service::RPCService& rpc_service,
@@ -49,9 +49,9 @@ VIUnsubscribeVehicleDataRequest::VIUnsubscribeVehicleDataRequest(
                    hmi_capabilities,
                    policy_handle) {}
 
-VIUnsubscribeVehicleDataRequest::~VIUnsubscribeVehicleDataRequest() {}
+VIGetDTCsRequest::~VIGetDTCsRequest() {}
 
-void VIUnsubscribeVehicleDataRequest::Run() {
+void VIGetDTCsRequest::Run() {
   LOG4CXX_AUTO_TRACE(logger_);
 
   SendRequest();
